backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncmy==0.2.8
pydantic-settings==2.3.4
httpx==0.27.0
python-telegram-bot==21.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
psutil==5.9.8
